Split up CompanyAi gc's into headers
