+++ src/OpenLoco/CMakeLists.txt
@@ -88,6 +88,7 @@ set(OLOCO_SOURCES
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/Company/RenameCompanyName.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/Company/RenameCompanyOwner.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/Company/UpdateOwnerStatus.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/CompanyAi/AiTrackReplacement.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/Docks/CreatePort.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/Docks/RemovePort.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/GameCommands.cpp"
@@ -443,6 +444,9 @@ set(OLOCO_HEADERS
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/Company/RenameCompanyName.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/Company/RenameCompanyOwner.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/Company/UpdateOwnerStatus.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/CompanyAi/AiTrackReplacement.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/CompanyAi/AiCreateTrackAndStation.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/CompanyAi/AiCreateRoadAndStation.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/Docks/CreatePort.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/Docks/RemovePort.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/GameCommands.h"
