+++ src/OpenLoco/src/World/CompanyAi/CompanyAi.cpp
@@ -6,6 +6,9 @@
 #include "GameCommands/Airports/RemoveAirport.h"
 #include "GameCommands/Company/BuildCompanyHeadquarters.h"
 #include "GameCommands/Company/RemoveCompanyHeadquarters.h"
+#include "GameCommands/CompanyAi/AiCreateRoadAndStation.h"
+#include "GameCommands/CompanyAi/AiCreateTrackAndStation.h"
+#include "GameCommands/CompanyAi/AiTrackReplacement.h"
 #include "GameCommands/Docks/CreatePort.h"
 #include "GameCommands/Docks/RemovePort.h"
 #include "GameCommands/GameCommands.h"
